pyproject.toml
[project]
name = "google-adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.3.0",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "requests>=2.32.4",
]
